+++ Abstractions/Models/Storage/LayoutUnit.cs
@@ -4,12 +4,12 @@ namespace Filuet.Hardware.Dispensers.Abstractions.Models
 {
     public abstract class LayoutUnit : ILayoutUnit
     {
-        public uint Number { get; private set; }
+        public ushort Number { get; private set; }
 
         public bool IsActive { get; private set; } = true;
 
         public void SetActive(bool active) => IsActive = active;
 
-        public void SetNumber(uint number) => Number = number;
+        public void SetNumber(ushort number) => Number = number;
     }
 }
